pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paid-telegram-bot"
version = "0.1.0"
description = "Enterprise Telegram bot management MCP server — users, payments, analytics, and 40+ tools for any AI agent."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "netflypsb"},
]
keywords = [
    "mcp", "telegram", "bot", "ai-agent", "enterprise",
    "payments", "stripe", "subscriptions", "analytics",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "mcp>=1.0.0",
    "python-telegram-bot>=21.0",
    "aiosqlite>=0.19.0",
    "mcp-marketplace-license>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/netflypsb/paid-telegram-bot-mcp"
Repository = "https://github.com/netflypsb/paid-telegram-bot-mcp"
Documentation = "https://github.com/netflypsb/paid-telegram-bot-mcp#readme"

[project.scripts]
paid-telegram-bot = "paid_telegram_bot:main"

[tool.hatch.build.targets.wheel]
packages = ["src/paid_telegram_bot"]
